Room Management working, Game nearly starting, just added Players class

+++ Assets/Scripts/Network/ANetworkView.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public abstract class ANetworkView : MonoBehaviour
 {
     [SerializeField]
-    private int viewId;
-    public int ViewId
+    private short viewId;
+    public short ViewId
     {
         get
         {
